[Trivial] A few cosmetic changes (#244)
- add `blockNumber` to MinimalTxData struct.
- a couple of nit picky log changes/additions 
- change let to const
- introduce utility method to convert `txHash --> MinimalTxData` and
utilize this method in unit tests.
